+++ src/vidlib/puregl2/vidlib_win32.c
@@ -248,7 +248,7 @@ VideoWindowProc (HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
 				INDEX idx;
 				struct dropped_file_acceptor_tag *callback;
 				//uint32_t namelen = DragQueryFile( hDrop, iFIle, NULL, 0 );
-				DragQueryFile( hDrop, iFile, buffer, (UINT)sizeof( buffer ) );
+				DragQueryFile( hDrop, (UINT)iFile, buffer, (UINT)sizeof( buffer ) );
 				//lprintf( "Accepting file drop [%s]", buffer );
 				LIST_FORALL( hVideo->dropped_file_acceptors, idx, struct dropped_file_acceptor_tag*, callback )
 				{
